refactor: Separate errors & cleanup imports
