+++ judger/src/client.rs
@@ -1,6 +1,9 @@
 pub mod config;
+mod err;
 pub mod model;
 pub mod sink;
+
+pub use err::*;
 
 use crate::{
     client::model::JobResultKind,
@@ -11,105 +14,20 @@ use crate::{
     tester::model::TestSuiteOptions,
 };
 use anyhow::Result;
-use async_trait::async_trait;
 use config::SharedClientData;
 
-use err_derive::Error;
-use futures::{
-    stream::{SplitSink, SplitStream},
-    Sink, SinkExt, StreamExt,
-};
+use futures::StreamExt;
 use http::Method;
 use model::*;
-use serde::Serialize;
 use serde_json::from_slice;
 use sink::*;
-use std::{collections::HashMap, fmt::Debug, path::PathBuf, sync::atomic::Ordering, sync::Arc};
-use tokio::{net::TcpStream, sync::Mutex};
-use tokio_tungstenite::{connect_async, tungstenite, MaybeTlsStream, WebSocketStream};
+use std::{collections::HashMap, path::PathBuf, sync::atomic::Ordering, sync::Arc};
+use tokio_tungstenite::{connect_async, tungstenite};
 use tracing::info_span;
 use tracing_futures::Instrument;
 use tungstenite::Message;
 
 // Arc<Mutex<WsSink>>==>>Arc<WsSink>
-
-#[derive(Debug, Error)]
-pub enum JobExecErr {
-    #[error(display = "No such file: {}", _0)]
-    NoSuchFile(String),
-
-    #[error(display = "No such config: {}", _0)]
-    NoSuchConfig(String),
-
-    #[error(display = "IO error: {}", _0)]
-    Io(#[error(source)] std::io::Error),
-
-    #[error(display = "Web request error: {}", _0)]
-    Request(#[error(source)] reqwest::Error),
-
-    #[error(display = "Websocket error: {}", _0)]
-    Ws(#[error(source, no_from)] tungstenite::Error),
-
-    #[error(display = "JSON error: {}", _0)]
-    Json(#[error(source)] serde_json::Error),
-
-    #[error(display = "TOML error: {}", _0)]
-    TomlDes(#[error(source)] toml::de::Error),
-
-    #[error(display = "Build error: {}", _0)]
-    Build(#[error(source)] crate::tester::BuildError),
-
-    #[error(display = "Execution error: {}", _0)]
-    Exec(#[error(source)] crate::tester::ExecError),
-
-    #[error(display = "Job was cancelled")]
-    Cancelled,
-
-    #[error(display = "Other error: {}", _0)]
-    Any(anyhow::Error),
-}
-
-impl From<tungstenite::error::Error> for JobExecErr {
-    fn from(e: tungstenite::error::Error) -> Self {
-        match e {
-            tungstenite::Error::Io(e) => JobExecErr::Io(e),
-            _ => JobExecErr::Ws(e),
-        }
-    }
-}
-
-macro_rules! anyhow_downcast_chain {
-    ($e:expr, $($ty:ty),*) => {
-        $(if $e.is::<$ty>(){
-            let e = $e.downcast::<$ty>().unwrap();
-            return e.into();
-        })*
-    };
-}
-
-impl From<anyhow::Error> for JobExecErr {
-    fn from(e: anyhow::Error) -> Self {
-        anyhow_downcast_chain!(
-            e,
-            std::io::Error,
-            crate::tester::BuildError,
-            crate::tester::ExecError,
-            toml::de::Error,
-            reqwest::Error
-        );
-        JobExecErr::Any(e)
-    }
-}
-
-#[derive(Debug, Error)]
-pub enum ClientConnectionErr {
-    #[error(display = "Websocket error: {}", _0)]
-    Ws(#[error(from)] tungstenite::Error),
-    #[error(display = "Bad access token")]
-    BadAccessToken,
-    #[error(display = "Bad register token")]
-    BadRegisterToken,
-}
 
 /// Try to register at the coordinator if no access token was specified.
 ///
