+++ judger/src/tester/model.rs
@@ -1,10 +1,8 @@
-use super::runner::CommandRunner;
 use anyhow::Result;
-use bollard::models::{BuildInfo, Mount};
+use bollard::models::Mount;
 use path_absolutize::Absolutize;
-use serde::de::Visitor;
-use serde::{self, Deserialize, Deserializer, Serialize};
-use std::{collections::HashMap, io, path::PathBuf, string::String, sync::Arc};
+use serde::{self, Deserialize, Serialize};
+use std::{collections::HashMap, path::PathBuf, string::String};
 use std::{path::Path, str::FromStr};
 
 /// A Host-to-container volume binding for the container.
